requirements.txt
telethon==1.35.0
pillow==10.4.0
requests==2.32.3
